refactor(cmd): split main.go into root.go and check.go
Move rootCmd to root.go and checkCmd/flags/init to check.go so that
main.go only contains the entry point.
